refactor: separando logica en clases
